requirements.txt
asgiref==3.7.2
attrs==23.2.0
channels==4.1.0
channels-redis==4.2.0
Django==5.0.3
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
djangochannelsrestframework==1.2.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
et-xmlfile==1.1.0
greenlet==3.0.3
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.8
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
